fix(client): Rename `components` directory
