+++ apps/client/src/components/Footer/Footer.tsx
