+++ apps/client/src/components/Footer/index.ts
