+++ apps/client/src/components/Termometer/Termometer.tsx
