+++ apps/client/src/components/Termometer/index.ts
